[FEAT]#30: Usecase로 로그인하면 저장해요
