+++ domain/src/main/kotlin/com/bff/wespot/domain/util/DataStoreKey.kt
@@ -1,0 +1,6 @@
+package com.bff.wespot.domain.util
+
+object DataStoreKey {
+    const val ACCESS_TOKEN = "access_token"
+    const val REFRESH_TOKEN = "refresh_token"
+}
